+++ src/app/admin/components/manage-products/manage-products.component.ts
@@ -11,5 +11,4 @@ export class ManageProductsComponent implements OnInit {
 
     ngOnInit(): void {
     }
-
 }