+++ src/app/core/models/AppSettings/app-settings.model.ts
@@ -1,0 +1,5 @@
+import { AppTheme } from '..';
+
+export interface AppSettingsModel {
+    THEME: AppTheme;
+}